Catch exception when game ends
The async generator in the server was raising an exception when the game
async generator raised its own exception to stop the polling loop.
That exception could be safely ignored in a python script, but it would
hang the code when inside a docker container. Thus this PR removes the bug
that was preventing the dockerized deployment of the game.